+++ app/src/main/java/com/example/guanguannfc/model/Dao/DaoBoxContent.java
@@ -22,7 +22,7 @@ public class DaoBoxContent {
     }
 
     //插入一条盒子物品数据：需要给定盒子ID、物品名称、物品数量
-    public void insert(Integer box_ID, String thing_name, Integer thing_num) {
+    public void insert(long box_ID, String thing_name, Integer thing_num) {
         SQLiteDatabase db = mDataBaseHelper.getWritableDatabase();
         String sql = "insert into " + GuanContract.BoxContent.TABLE_NAME + "(box_ID,thing_name,thing_num,created_time,updated_time) values(?,?,?,?,?)";
 
@@ -34,7 +34,7 @@ public class DaoBoxContent {
     }
 
     //删除某个盒子的所有物品：需要给定盒子ID
-    public void delete(Integer box_ID) {
+    public void delete(long box_ID) {
         SQLiteDatabase db = mDataBaseHelper.getWritableDatabase();
         String sql = "delete from " + GuanContract.BoxContent.TABLE_NAME + " where box_ID=?";
         db.execSQL(sql, new Object[]{box_ID});
@@ -42,7 +42,7 @@ public class DaoBoxContent {
     }
 
     //删除某个盒子中某个物品：需要给定盒子ID、物品名称
-    public void delete(Integer box_ID, String thing_name) {
+    public void delete(long box_ID, String thing_name) {
         SQLiteDatabase db = mDataBaseHelper.getWritableDatabase();
         String sql = "delete from " + GuanContract.BoxContent.TABLE_NAME + " where box_ID=? and thing_name=?";
         db.execSQL(sql, new Object[]{box_ID, thing_name});
@@ -50,7 +50,7 @@ public class DaoBoxContent {
     }
 
     //更新某个盒子中某个物品的数量：需要给定盒子的ID、物品的名称、更新后物品的数量
-    public void update(Integer box_ID, String thing_name, Integer thing_num) {
+    public void update(long box_ID, String thing_name, Integer thing_num) {
         SQLiteDatabase db = mDataBaseHelper.getWritableDatabase();
         String sql = "update " + GuanContract.BoxContent.TABLE_NAME + " set thing_num=? , updated_time=? where box_ID=? and thing_name=?";
         Date date = new Date();
@@ -61,7 +61,7 @@ public class DaoBoxContent {
     }
 
     //指定盒子内物品名称查重，存在返回true，反之false：需给定盒子ID、查重物品名称
-    public boolean loadQuery(Integer box_ID, String thing_name) {
+    public boolean loadQuery(long box_ID, String thing_name) {
         SQLiteDatabase db = mDataBaseHelper.getWritableDatabase();
         String sql = "select * from " + GuanContract.BoxContent.TABLE_NAME + " where box_ID=? and thing_name=?";
         Cursor cursor = db.rawQuery(sql,new String[]{String.valueOf(box_ID),thing_name});
